Update README.md
Fill the field from the input and print it as in the previous stage.
 Find the state in which the game is at the moment and print it. Possible states:
"Game not finished" - when no side has a three in a row but the field has empty cells;
"Draw" - when no side has a three in a row and the field has no empty cells;
"X wins" - when the field has three X in a row;
"O wins" - when the field has three O in a row;
"Impossible" - when the field has three X in a row as well as three O in a row. Or the field has a lot more X's that O's or vice versa (if the difference is 2 or more, should be 1 or 0). For this stage, consider that the game can be started both as X's or as O's. 
Also, you can use ' ' or '_' to print empty cells - it's up to you.